rename enums to categoryStringDefinitions.ts and move to utils
