+++ api/src/utils/categoryStringDefinitions.ts
